M2-5565: M2-5566 DataViz changes
