+++ src/modules/Dashboard/features/RespondentData/RespondentData.types.ts
@@ -85,17 +85,3 @@ export type RespondentsDataFormValues = {
   subscalesFrequency: number;
   responseDate: null | Date;
 };
-
-export type GetIdentifiersVersions = {
-  activity: DatavizActivity;
-};
-
-export type FetchAnswers = {
-  activity: DatavizActivity;
-  endDate?: Date;
-  startTime?: string;
-  endTime?: string;
-  filterByIdentifier?: boolean;
-  identifier?: AutocompleteOption[];
-  versions?: AutocompleteOption[];
-};
