+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/Report/Report.const.ts
@@ -3,4 +3,3 @@ export const LATEST_REPORT_DEFAULT_NAME = 'Report.pdf';
 export const LATEST_REPORT_REGEX = /filename="(?<filename>.*)"/;
 
 export const DEFAULT_DATE_MIN = new Date(1970, 0, 1, 23, 0, 0);
-export const DEFAULT_DATE_MAX = new Date(1970, 0, 2, 23, 59, 0);
