+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/RespondentDataSummary.const.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_DATE_MAX = new Date(1970, 0, 2, 23, 59, 0);
