+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/RespondentDataSummary.types.ts
@@ -1,0 +1,16 @@
+import { DatavizActivity } from 'api';
+import { AutocompleteOption } from 'shared/components/FormComponents';
+
+export type FetchAnswers = {
+  activity: DatavizActivity;
+  endDate?: Date;
+  startTime?: string;
+  endTime?: string;
+  filterByIdentifier?: boolean;
+  identifier?: AutocompleteOption[];
+  versions?: AutocompleteOption[];
+};
+
+export type GetIdentifiersVersions = {
+  activity: DatavizActivity;
+};
